requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-filter==24.2
django-extensions==3.2.3
gunicorn==22.0.0
pylint==3.1.0
pylint-django==2.5.5
autopep8==2.1.0
nose==1.3.7
requests==2.32.2
responses==0.25.0
ipython==8.24.0
tqdm==4.66.4
python-decouple==3.8
psycopg2-binary==2.9.9
whitenoise==6.7.0
redis==5.0.3
sentry-sdk==1.43.0
celery==5.3.6
django-celery-beat==2.6.0
faker==24.4.0
django-silk==5.1.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
channels==4.0.0
daphne==4.1.0
channels_redis==4.2.0
clickhouse-connect==0.7.9
flower==2.0.1
uvicorn==0.30.1
websockets==12.0
# asgiref==3.7.2
py-spy==0.3.14
memory_profiler==0.61.0
jsonpickle==3.0.3
django-celery-results==2.5.1
ndg-httpsclient==0.5.1
pyopenssl==24.1.0
pyasn1==0.5.1
redlock==1.2.0
django-db-connection-retrier==1.0
psycopg==3.2.9
psycopg_binary==3.2.9
